상품 엔드포인트 수정 완료
