+++ src/main/java/edu/example/dev_2_cc/api_controller/MypageController.java
@@ -1,0 +1,13 @@
+package edu.example.dev_2_cc.api_controller;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.log4j.Log4j2;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+@Log4j2
+@RequestMapping("/cc/mypage")
+public class MypageController {
+}
